+++ R/TorchOptimizer.R
@@ -262,8 +262,8 @@ mlr3torch_optimizers$add("adamw",
     p = ps(
       lr           = p_dbl(default = 0.001, lower = 0, upper = Inf, tags = "train"),
       betas        = p_uty(default = c(0.9, 0.999), tags = "train", custom_check = check_betas),
-      eps          = p_dbl(default = 1e-08, lower = 1e-16, upper = 1e-4, tags = "train"),
-      weight_decay = p_dbl(default = 0.01, lower = 0, upper = 1, tags = "train"),
+      eps          = p_dbl(default = 1e-08, lower = 1e-16, tags = "train"),
+      weight_decay = p_dbl(default = 0.01, lower = 0, tags = "train"),
       amsgrad      = p_lgl(default = FALSE, tags = "train")
     )
     TorchOptimizer$new(
@@ -288,8 +288,8 @@ mlr3torch_optimizers$add("adam",
     p = ps(
       lr           = p_dbl(default = 0.001, lower = 0, upper = Inf, tags = "train"),
       betas        = p_uty(default = c(0.9, 0.999), tags = "train", custom_check = check_betas),
-      eps          = p_dbl(default = 1e-08, lower = 1e-16, upper = 1e-4, tags = "train"),
-      weight_decay = p_dbl(default = 0, lower = 0, upper = 1, tags = "train"),
+      eps          = p_dbl(default = 1e-08, lower = 1e-16, tags = "train"),
+      weight_decay = p_dbl(default = 0, lower = 0, tags = "train"),
       amsgrad      = p_lgl(default = FALSE, tags = "train")
     )
     TorchOptimizer$new(
@@ -308,7 +308,7 @@ mlr3torch_optimizers$add("sgd",
       lr           = p_dbl(lower = 0, tags = c("required", "train")),
       momentum     = p_dbl(0, 1, default = 0, tags = "train"),
       dampening    = p_dbl(default = 0, lower = 0, upper = 1, tags = "train"),
-      weight_decay = p_dbl(0, 1, default = 0, tags = "train"),
+      weight_decay = p_dbl(lower = 0, default = 0, tags = "train"),
       nesterov     = p_lgl(default = FALSE, tags = "train")
     )
     TorchOptimizer$new(
@@ -327,7 +327,7 @@ mlr3torch_optimizers$add("rmsprop",
       lr           = p_dbl(default = 0.01, lower = 0, tags = "train"),
       alpha        = p_dbl(default = 0.99, lower = 0, upper = 1, tags = "train"),
       eps          = p_dbl(default = 1e-08, lower = 0, upper = Inf, tags = "train"),
-      weight_decay = p_dbl(default = 0, lower = 0, upper = 1, tags = "train"),
+      weight_decay = p_dbl(default = 0, lower = 0, tags = "train"),
       momentum     = p_dbl(default = 0, lower = 0, upper = 1, tags = "train"),
       centered     = p_lgl(default = FALSE, tags = "train")
     )
@@ -348,7 +348,7 @@ mlr3torch_optimizers$add("adagrad",
       lr_decay                  = p_dbl(default = 0, lower = 0, upper = 1, tags = "train"),
       weight_decay              = p_dbl(default = 0, lower = 0, upper = 1, tags = "train"),
       initial_accumulator_value = p_dbl(default = 0, lower = 0, tags = "train"),
-      eps                       = p_dbl(default = 1e-10, lower = 1e-16, upper = 1e-4, tags = "train")
+      eps                       = p_dbl(default = 1e-10, lower = 1e-16, tags = "train")
     )
     TorchOptimizer$new(
       torch_optimizer = torch::optim_ignite_adagrad,